+++ pythonFiles/tests/ipython/scripts.py
@@ -75,7 +75,9 @@ def get_data_frame_info(variables, name, capsys):
     varJson = find_variable_json(variables, name)
     path = os.path.dirname(os.path.abspath(__file__))
     file = os.path.abspath(
-        os.path.join(path, "../../datascience/getJupyterVariableDataFrameInfo.py")
+        os.path.join(
+            path, "../../datascience_modules/getJupyterVariableDataFrameInfo.py"
+        )
     )
     keys = dict([("_VSCode_JupyterTestValue", json.dumps(varJson))])
     if execute_script(file, keys):
@@ -88,7 +90,9 @@ def get_data_frame_info(variables, name, capsys):
 def get_data_frame_rows(varJson, start, end, capsys):
     path = os.path.dirname(os.path.abspath(__file__))
     file = os.path.abspath(
-        os.path.join(path, "../../datascience/getJupyterVariableDataFrameRows.py")
+        os.path.join(
+            path, "../../datascience_modules/getJupyterVariableDataFrameRows.py"
+        )
     )
     keys = dict(
         [
