+++ src/client/datascience/jupyter/interpreter/jupyterCommand.ts
@@ -107,7 +107,14 @@ class InterpreterJupyterCommand implements IJupyterCommand {
                 ) {
                     try {
                         const output = await svc.exec(
-                            [path.join(EXTENSION_ROOT_DIR, 'pythonFiles', 'datascience', 'jupyter_nbInstalled.py')],
+                            [
+                                path.join(
+                                    EXTENSION_ROOT_DIR,
+                                    'pythonFiles',
+                                    'datascience_modules',
+                                    'jupyter_nbInstalled.py'
+                                )
+                            ],
                             {}
                         );
                         if (output.stdout.toLowerCase().includes('available')) {
@@ -264,7 +271,7 @@ export class InterpreterJupyterKernelSpecCommand extends InterpreterJupyterComma
             bypassCondaExecution: true
         });
         return activatedEnv.exec(
-            [path.join(EXTENSION_ROOT_DIR, 'pythonFiles', 'datascience', 'getJupyterKernels.py')],
+            [path.join(EXTENSION_ROOT_DIR, 'pythonFiles', 'datascience_modules', 'getJupyterKernels.py')],
             { ...options, throwOnStdErr: true }
         );
     }
@@ -275,7 +282,7 @@ export class InterpreterJupyterKernelSpecCommand extends InterpreterJupyterComma
             bypassCondaExecution: true
         });
         return activatedEnv.exec(
-            [path.join(EXTENSION_ROOT_DIR, 'pythonFiles', 'datascience', 'getJupyterKernelspecVersion.py')],
+            [path.join(EXTENSION_ROOT_DIR, 'pythonFiles', 'datascience_modules', 'getJupyterKernelspecVersion.py')],
             { ...options, throwOnStdErr: true }
         );
     }
